update exos de jeudi

+++ app/controllers/gossips_controller.rb
@@ -1,15 +1,24 @@
 class GossipsController < ApplicationController
 
+ before_action :authenticate_user, only: [:show, :new, :create, :edit, :update, :destroy]
+
+
+
   def index
     @gossip = Gossip.all
     @user = User.all
   end
+
+
+
 
   def show
     @gossip = Gossip.find(params[:id])
     @user = User.all
 
   end
+
+
 
   def new
     # Méthode qui crée un potin vide et l'envoie à une view qui affiche le formulaire pour 'le remplir' (new.html.erb)
@@ -21,9 +30,8 @@ class GossipsController < ApplicationController
     # Méthode qui créé un potin à partir du contenu du formulaircite de new.html.erb, soumis par l'utilisateur
     # pour info, le contenu de ce formulaire sera accessible dans le hash params (ton meilleur pote)
     # Une fois la création faite, on redirige généralement vers la méthode show (pour afficher le potin créé)
-  #anonymous=User.new(first_name:"Anonymous", last_name:"unknown",age: 99,description: "unknown",email:"unknown#{rand(1000)}@mail.com",city_id:City.all.sample.id)
-  #anonymous.save
-  @gossip = Gossip.create(title: params[:title], content: params[:content], user_id: 2) # avec xxx qui sont les données obtenues à partir du formulaire
+ 
+  @gossip = Gossip.create(title: params[:title], content: params[:content], user_id: session[:user_id]) # avec xxx qui sont les données obtenues à partir du formulaire
 
   if @gossip.save # essaie de sauvegarder en base @gossip
     # si ça marche, il redirige vers la page d'index du site
@@ -37,24 +45,31 @@ class GossipsController < ApplicationController
   end
 
 
+
   def edit
     # Méthode qui récupère le potin concerné et l'envoie à la view edit (edit.html.erb) pour affichage dans un formulaire d'édition
   @gossip = Gossip.find(params[:id])
   end
 
+
+
   def update
-    
+
      @gossip = Gossip.find(params[:id])
      post_params = params.permit(:title, :content)
 
-     if @gossip.update(post_params, user_id: 2) #title: params[:title], content: params[:content]
+  unless compare_user(session[:user_id], @gossip.user_id) == false
+     if @gossip.update(title: params[:title], content: params[:content])
      @gossips=Gossip.all
      redirect_to gossips_path
 
      else
      render :edit
-  
-end
+  end  
+     end
+     redirect_to gossips_path
+
+
 end
 
 
@@ -65,9 +80,20 @@ end
 
     @gossip = Gossip.find(params[:id])
 
-   if  @gossip.destroy
+   unless compare_user(session[:user_id], @gossip.user_id) == false
+      if  @gossip.destroy
        redirect_to gossips_path 
-    else render :show
+       else 
+        render :show
   end
+end 
+  
 end
+
+
+
 end
+
+
+
+
